folder structure : placed the Subtotal css and js files at right place
